+++ zellij-utils/src/data.rs
@@ -3461,6 +3461,12 @@ pub enum PluginCommand {
     OpenCommandPaneInPlaceOfPaneId(PaneId, CommandToRun, bool, Context), // bool = close_replaced_pane
     OpenTerminalPaneInPlaceOfPaneId(PaneId, FileToOpen, bool),
     OpenEditPaneInPlaceOfPaneId(PaneId, FileToOpen, bool, Context),
+    HideFloatingPanes {
+        tab_id: Option<usize>,
+    },
+    ShowFloatingPanes {
+        tab_id: Option<usize>,
+    },
 }
 
 // Response type for plugin API methods that open a pane in a new tab